refac(pdfid): returned to classic pdfid, from now on we'll be using official pdfid utility functions
