+++ main.py
@@ -1,7 +1,8 @@
 import argparse
-import subprocess
 import sys
 from pathlib import Path
+
+from pdfid.pdfid import PDFiD
 
 
 class CustomArgumentParser(argparse.ArgumentParser):
@@ -49,9 +50,15 @@ parser.add_argument(
 # Parsed arguments
 args = parser.parse_args()
 
+directory_path_arg = ""
+if args.directory:
+    directory_path_arg = str(Path.home() / args.directory)
+
+file_path_arg = ""
+if args.file:
+    file_path_arg = str(Path.home() / args.file)
+
 level_arg = args.level
-directory_path_arg = args.directory
-file_path_arg = args.file
 watch_arg = args.watch
 
 
@@ -63,63 +70,16 @@ if (
     parser.error(
         "incompatible options: file & directory or watch can't be used at the same time"
     )
+elif not file_path_arg and not directory_path_arg:
+    parser.error("atmost one of the options -f or -d must be defined")
 
-file_path = Path.home() / file_path_arg
-command = subprocess.run(
-    ["python3", "pdfid", file_path], capture_output=True, text=True
-)
-
-
-def pretty_output(head=False):
-    data_lines = command.stdout.splitlines()
-    if not head:
-        data_lines = data_lines[:2]
-
-    # Pretty subprocess (pdfid) output
-    keyword_counts = {}
-    for line in data_lines:
-        # Strip whitespace
-        line = line.strip()
-        if not line:
-            continue
-        # Split by whitespace, maxsplit=1 to separate key and value
-        parts = line.split(maxsplit=1)
-        if len(parts) == 2:
-            key, value = parts
-            try:
-                keyword_counts[key] = int(value)
-            except ValueError:
-                # Not a number, ignore or handle differently
-                pass
-
-    return keyword_counts
-
-
-# Level 1
-def level_1(dict):
-    suspicious_keys = ["/JS", "/JavaScript", "/AA", "/OpenAction"]
-    threats_found = {}
-    threats = 0
-
-    print("Hunting Basic Level 1 Threats...\n")
-    for key in suspicious_keys:
-        count = dict.get(key, 0)
-        if count > 0:
-            threats_found[key] = count
-            threats += 1
-        print(f"{key}: {count}")
-    print("\n")
-
-    if threats_found:
-        print("Suspicous Items Found:\n")
-
-
+result = None
 if level_arg == 1:
-    keyword_counts = pretty_output()
-    level_1(keyword_counts)
+    keyword_counts = PDFiD(file_path_arg)
+    result = keyword_counts
 
-# for node in command.getElementsByTagName("Keyword"):
-#     if node.getAttribute("Name") in ["/JS", "/JavaScript", "/AA", "/OpenAction"]:
-#         name = node.getAttribute("Name")
-#         count = node.getAttribute("Count")
-#         print(f"{name}: {count}")
+for node in result.getElementsByTagName("Keyword"):
+    if node.getAttribute("Name") in ["/JS", "/JavaScript", "/AA", "/OpenAction"]:
+        name = node.getAttribute("Name")
+        count = node.getAttribute("Count")
+        print(f"{name}: {count}")
